+++ pdf/lib/widgets.dart
@@ -41,6 +41,7 @@ part 'widgets/image.dart';
 part 'widgets/multi_page.dart';
 part 'widgets/page.dart';
 part 'widgets/page_theme.dart';
+part 'widgets/partitions.dart';
 part 'widgets/placeholders.dart';
 part 'widgets/progress.dart';
 part 'widgets/qrcode.dart';