+++ pdf/test/all_tests.dart
@@ -35,6 +35,7 @@ import 'widget_container_test.dart' as widget_container;
 import 'widget_flex_test.dart' as widget_flex;
 import 'widget_grid_view_test.dart' as widget_grid_view;
 import 'widget_multipage_test.dart' as widget_multipage;
+import 'widget_partitions_test.dart' as widget_partitions;
 import 'widget_table_test.dart' as widget_table;
 import 'widget_test.dart' as widget;
 import 'widget_text_test.dart' as widget_text;
@@ -62,6 +63,7 @@ void main() {
   widget_flex.main();
   widget_grid_view.main();
   widget_multipage.main();
+  widget_partitions.main();
   widget_table.main();
   widget_text.main();
   widget_theme.main();